Split modals in lecturer course page into files
