+++ app/courses/[courseId]/lecturer.tsx
@@ -1,59 +1,26 @@
 "use client";
 import { lecturerLecturerCourseIdQuery } from "@/__generated__/lecturerLecturerCourseIdQuery.graphql";
-import {
-  Alert,
-  Backdrop,
-  Button,
-  CircularProgress,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
-  IconButton,
-  Switch,
-  TextField,
-  Typography,
-} from "@mui/material";
+import { Button, IconButton, Typography } from "@mui/material";
 import Error from "next/error";
 import { useParams } from "next/navigation";
-import {
-  graphql,
-  useFragment,
-  useLazyLoadQuery,
-  useMutation,
-} from "react-relay";
+import { graphql, useLazyLoadQuery } from "react-relay";
 
-import { ContentType } from "@/__generated__/MediaContentModalCreateMutation.graphql";
 import { MediaRecordSelector$key } from "@/__generated__/MediaRecordSelector.graphql";
-import { lecturerCreateFlashcardAssessmentMutation } from "@/__generated__/lecturerCreateFlashcardAssessmentMutation.graphql";
-import { lecturerCreateFlashcardSetMutation } from "@/__generated__/lecturerCreateFlashcardSetMutation.graphql";
-import { lecturerEditCourseAddChapterModalFragment$key } from "@/__generated__/lecturerEditCourseAddChapterModalFragment.graphql";
-import { lecturerEditCourseChaptersMutation } from "@/__generated__/lecturerEditCourseChaptersMutation.graphql";
-import { lecturerEditCourseEditChapterModalFragment$key } from "@/__generated__/lecturerEditCourseEditChapterModalFragment.graphql";
-import { lecturerEditCourseEditChaptersMutation } from "@/__generated__/lecturerEditCourseEditChaptersMutation.graphql";
-import { lecturerEditCourseGeneralFragment$key } from "@/__generated__/lecturerEditCourseGeneralFragment.graphql";
-import { lecturerEditCourseMutation } from "@/__generated__/lecturerEditCourseMutation.graphql";
-import { SkillType } from "@/__generated__/lecturerEditFlashcardSetModalFragment.graphql";
-import {
-  AssessmentMetadataFormSection,
-  AssessmentMetadataPayload,
-} from "@/components/AssessmentMetadataFormSection";
 import { ChapterContent } from "@/components/ChapterContent";
 import { ChapterHeader } from "@/components/ChapterHeader";
 import { Content, ContentLink, ProgressFrame } from "@/components/Content";
-import {
-  ContentMetadataFormSection,
-  ContentMetadataPayload,
-} from "@/components/ContentMetadataFormSection";
-import { Form, FormSection } from "@/components/Form";
-import { Add, Edit, RemoveRedEye, Settings } from "@mui/icons-material";
-import { DatePicker } from "@mui/x-date-pickers";
-import dayjs, { Dayjs } from "dayjs";
+import { Add, RemoveRedEye, Settings } from "@mui/icons-material";
+import dayjs from "dayjs";
 import { orderBy } from "lodash";
 import { useState } from "react";
 import { MediaContentModal } from "../../../components/MediaContentModal";
 import { Topic, TopicContent } from "@/components/Topic";
 import { TopicStage } from "@/components/TopicStage";
+import { AddFlashcardSetModal } from "@/components/AddFlashcardSetModal";
+import { AddChapterModal } from "@/components/AddChapterModal";
+import { EditChapterModal } from "@/components/EditChapterModal";
+import { EditCourseModal } from "@/components/EditCourseModal";
+import { AddFlashcardSetModalFragment$key } from "@/__generated__/AddFlashcardSetModalFragment.graphql";
 
 export default function LecturerCoursePage() {
   // Get course id from url
@@ -62,11 +29,6 @@ export default function LecturerCoursePage() {
 
   // Info dialog
   const [infoDialogOpen, setInfoDialogOpen] = useState(false);
-
-  // Flashcard set dialog
-  const [addFlashcardsChapter, setAddFlashcardsChapter] = useState<
-    string | null
-  >(null);
 
   // Fetch course data
   const { coursesById, ...query } =
@@ -78,12 +40,13 @@ export default function LecturerCoursePage() {
           coursesById(ids: $id) {
             title
             description
-            ...lecturerEditCourseAddChapterModalFragment
-            ...lecturerEditCourseGeneralFragment
+            ...AddChapterModalFragment
+            ...EditCourseModalFragment
             chapters {
               elements {
                 __id
-                ...lecturerEditCourseEditChapterModalFragment
+                ...EditChapterModalFragment
+                ...AddFlashcardSetModalFragment
                 id
                 title
                 number
@@ -132,7 +95,7 @@ export default function LecturerCoursePage() {
           <Settings />
         </IconButton>
       </div>
-      <EditGeneral
+      <EditCourseModal
         _course={course}
         open={infoDialogOpen}
         onClose={() => setInfoDialogOpen(false)}
@@ -191,24 +154,12 @@ export default function LecturerCoursePage() {
                     <ContentLink key={content.id} _content={content} />
                   ))}
                   <div className="mt-4 flex flex-col items-start">
-                    <Button
-                      startIcon={<Add />}
-                      onClick={() => setAddFlashcardsChapter(chapter.id)}
-                    >
-                      Add flashcards
-                    </Button>
+                    <AddFlashcardSetButton _chapter={chapter} />
                     <AddMediaButton
                       _mediaRecords={query}
                       chapterId={chapter.id}
                     />
                   </div>
-                  {chapter.id === addFlashcardsChapter && (
-                    <AddFlashcardSetModal
-                      onClose={() => setAddFlashcardsChapter(null)}
-                      chapterId={chapter.id}
-                      chapterRecordId={chapter.__id}
-                    />
-                  )}
                 </TopicStage>
               </TopicContent>
             </Topic>
@@ -216,563 +167,6 @@ export default function LecturerCoursePage() {
         </section>
       ))}
     </main>
-  );
-}
-
-function EditGeneral({
-  _course,
-  open,
-  onClose,
-}: {
-  open: boolean;
-  onClose: () => void;
-  _course: lecturerEditCourseGeneralFragment$key;
-}) {
-  const course = useFragment(
-    graphql`
-      fragment lecturerEditCourseGeneralFragment on Course {
-        id
-        title
-        description
-        startDate
-        endDate
-        published
-      }
-    `,
-    _course
-  );
-
-  const [updateCourse, isUpdating] =
-    useMutation<lecturerEditCourseMutation>(graphql`
-      mutation lecturerEditCourseMutation($course: UpdateCourseInput!) {
-        updateCourse(input: $course) {
-          id
-          ...lecturerEditCourseGeneralFragment
-        }
-      }
-    `);
-
-  const [title, setTitle] = useState(course.title);
-  const [description, setDescription] = useState(course.description);
-  const [startDate, setStartDate] = useState<Dayjs | null>(
-    dayjs(course.startDate)
-  );
-  const [endDate, setEndDate] = useState<Dayjs | null>(dayjs(course.endDate));
-  const [publish, setPublish] = useState(course.published);
-
-  const [error, setError] = useState<any>(null);
-  const valid =
-    title !== "" &&
-    startDate != null &&
-    startDate.isValid() &&
-    endDate != null &&
-    endDate.isValid();
-
-  function handleSubmit() {
-    updateCourse({
-      variables: {
-        course: {
-          id: course.id,
-          title,
-          description,
-          startDate: startDate!.toISOString(),
-          endDate: endDate!.toISOString(),
-          published: publish,
-        },
-      },
-      onError(error) {
-        setError(error);
-      },
-      onCompleted() {
-        onClose();
-      },
-    });
-  }
-
-  function handleReset() {
-    setTitle(course.title);
-    setDescription(course.description);
-    setStartDate(dayjs(course.startDate));
-    setEndDate(dayjs(course.endDate));
-    setPublish(course.published);
-    setError(null);
-  }
-
-  return (
-    <>
-      <Dialog maxWidth="lg" onClose={onClose} open={open}>
-        <DialogTitle>{title}</DialogTitle>
-        <DialogContent>
-          {error?.source.errors.map((err: any, i: number) => (
-            <Alert
-              key={i}
-              severity="error"
-              sx={{ minWidth: 400, maxWidth: 800, width: "fit-content" }}
-              onClose={() => setError(null)}
-            >
-              {err.message}
-            </Alert>
-          ))}
-          <Form>
-            <FormSection title="Course details">
-              <TextField
-                className="w-96"
-                label="Title"
-                variant="outlined"
-                value={title}
-                error={title === ""}
-                onChange={(e) => setTitle(e.target.value)}
-                required
-              />
-              <TextField
-                className="w-96"
-                label="Description"
-                variant="outlined"
-                value={description}
-                onChange={(e) => setDescription(e.target.value)}
-                multiline
-              />
-            </FormSection>
-
-            <FormSection title="Start and end">
-              <DatePicker
-                label="Start date"
-                value={startDate}
-                maxDate={endDate ?? undefined}
-                onChange={setStartDate}
-                slotProps={{
-                  textField: {
-                    required: true,
-                    error: startDate == null || !startDate.isValid(),
-                  },
-                }}
-              />
-              <DatePicker
-                label="End date"
-                value={endDate}
-                minDate={startDate ?? undefined}
-                defaultCalendarMonth={startDate ?? undefined}
-                onChange={setEndDate}
-                slotProps={{
-                  textField: {
-                    required: true,
-                    error: endDate == null || !endDate.isValid(),
-                  },
-                }}
-              />
-            </FormSection>
-
-            <FormSection title="Published">
-              <Switch
-                checked={publish}
-                onChange={(e) => setPublish(e.target.checked)}
-              />
-            </FormSection>
-          </Form>
-          <Backdrop open={isUpdating} sx={{ zIndex: "modal" }}>
-            <CircularProgress />
-          </Backdrop>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleReset}>Reset</Button>
-          <Button disabled={!valid} variant="contained" onClick={handleSubmit}>
-            Update
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </>
-  );
-}
-
-function EditChapterModal({
-  _chapter,
-}: {
-  _chapter: lecturerEditCourseEditChapterModalFragment$key;
-}) {
-  const [openModal, setOpenModal] = useState(false);
-
-  const chapter = useFragment(
-    graphql`
-      fragment lecturerEditCourseEditChapterModalFragment on Chapter {
-        id
-        title
-        description
-        startDate
-        endDate
-        suggestedStartDate
-        suggestedEndDate
-        course {
-          id
-        }
-        number
-      }
-    `,
-    _chapter
-  );
-
-  const handleOpenModal = () => {
-    setOpenModal(true);
-  };
-
-  const handleCloseModal = () => {
-    setOpenModal(false);
-    setTitle(chapter.title);
-    setDescription(chapter.description);
-    setStartDate(dayjs(chapter.startDate));
-    setEndDate(dayjs(chapter.endDate));
-    setSuggestedStartDate(dayjs(chapter.suggestedStartDate));
-    setSuggestedEndDate(dayjs(chapter.suggestedEndDate));
-  };
-
-  const [title, setTitle] = useState(chapter.title);
-  const [description, setDescription] = useState(chapter.description);
-  const [startDate, setStartDate] = useState<Dayjs | null>(
-    dayjs(chapter.startDate)
-  );
-  const [endDate, setEndDate] = useState<Dayjs | null>(dayjs(chapter.endDate));
-  const [suggestedStartDate, setSuggestedStartDate] = useState<Dayjs | null>(
-    dayjs(chapter.suggestedStartDate)
-  );
-  const [suggestedEndDate, setSuggestedEndDate] = useState<Dayjs | null>(
-    dayjs(chapter.suggestedEndDate)
-  );
-
-  const [error, setError] = useState<any>(null);
-  const valid =
-    title !== "" &&
-    startDate != null &&
-    startDate.isValid() &&
-    endDate != null &&
-    endDate.isValid();
-
-  const [updateChapter, isUpdating] =
-    useMutation<lecturerEditCourseEditChaptersMutation>(graphql`
-      mutation lecturerEditCourseEditChaptersMutation(
-        $chapter: UpdateChapterInput!
-      ) {
-        updateChapter(input: $chapter) {
-          id
-          ...lecturerEditCourseEditChapterModalFragment
-        }
-      }
-    `);
-
-  function handleSubmit() {
-    updateChapter({
-      variables: {
-        chapter: {
-          id: chapter.id,
-          title,
-          description,
-          startDate: startDate!.toISOString(),
-          endDate: endDate!.toISOString(),
-          suggestedStartDate: suggestedStartDate!.toISOString(),
-          suggestedEndDate: suggestedEndDate!.toISOString(),
-          number: chapter.number,
-        },
-      },
-      onCompleted() {
-        handleCloseModal();
-      },
-      onError(error) {
-        setError(error);
-      },
-    });
-  }
-
-  return (
-    <>
-      <div onClick={handleOpenModal} className="cursor-pointer">
-        <Edit className="text-gray-400 mb-1" fontSize="small" />
-      </div>
-      <Dialog maxWidth="md" open={openModal} onClose={handleCloseModal}>
-        <DialogTitle>Edit Chapter</DialogTitle>
-        <DialogContent>
-          {error?.source.errors.map((err: any, i: number) => (
-            <Alert key={i} severity="error" onClose={() => setError(null)}>
-              {err.message}
-            </Alert>
-          ))}
-          <Form>
-            <FormSection title="General">
-              <TextField
-                className="w-96"
-                label="Title"
-                variant="outlined"
-                value={title}
-                error={title === ""}
-                onChange={(e) => setTitle(e.target.value)}
-                required
-              />
-              <TextField
-                className="w-96"
-                label="Description"
-                variant="outlined"
-                value={description}
-                onChange={(e) => setDescription(e.target.value)}
-                multiline
-              />
-            </FormSection>
-
-            <FormSection title="Start and end">
-              <DatePicker
-                label="Start date"
-                value={startDate}
-                maxDate={endDate ?? undefined}
-                onChange={setStartDate}
-                slotProps={{
-                  textField: {
-                    required: true,
-                    error: startDate == null || !startDate.isValid(),
-                  },
-                }}
-              />
-              <DatePicker
-                label="End date"
-                value={endDate}
-                minDate={startDate ?? undefined}
-                defaultCalendarMonth={startDate ?? undefined}
-                onChange={setEndDate}
-                slotProps={{
-                  textField: {
-                    required: true,
-                    error: endDate == null || !endDate.isValid(),
-                  },
-                }}
-              />
-            </FormSection>
-            <FormSection title="Suggested start and end">
-              <DatePicker
-                label="Suggested start date"
-                value={suggestedStartDate}
-                maxDate={suggestedEndDate ?? undefined}
-                onChange={setSuggestedStartDate}
-                slotProps={{
-                  textField: {
-                    required: true,
-                    error:
-                      suggestedStartDate == null ||
-                      !suggestedStartDate.isValid(),
-                  },
-                }}
-              />
-              <DatePicker
-                label="Suggested end date"
-                value={suggestedEndDate}
-                minDate={suggestedStartDate ?? undefined}
-                defaultCalendarMonth={suggestedStartDate ?? undefined}
-                onChange={setSuggestedEndDate}
-                slotProps={{
-                  textField: {
-                    required: true,
-                    error:
-                      suggestedEndDate == null || !suggestedEndDate.isValid(),
-                  },
-                }}
-              />
-            </FormSection>
-          </Form>
-          <Backdrop open={isUpdating} sx={{ zIndex: "modal" }}>
-            <CircularProgress />
-          </Backdrop>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleCloseModal}>Cancel</Button>
-          <Button disabled={!valid} onClick={handleSubmit}>
-            Save
-          </Button>
-        </DialogActions>
-      </Dialog>
-      <Backdrop open={isUpdating} sx={{ zIndex: "modal" }}>
-        <CircularProgress />
-      </Backdrop>
-    </>
-  );
-}
-
-function AddChapterModal({
-  open,
-  onClose,
-  _course,
-}: {
-  open: boolean;
-  onClose: () => void;
-  _course: lecturerEditCourseAddChapterModalFragment$key;
-}) {
-  const course = useFragment(
-    graphql`
-      fragment lecturerEditCourseAddChapterModalFragment on Course {
-        id
-        chapters {
-          elements {
-            id
-            title
-          }
-        }
-      }
-    `,
-    _course
-  );
-
-  const [title, setTitle] = useState("");
-  const [description, setDescription] = useState("");
-  const [startDate, setStartDate] = useState<Dayjs | null>(dayjs());
-  const [endDate, setEndDate] = useState<Dayjs | null>(dayjs());
-  const [suggestedStartDate, setSuggestedStartDate] = useState<Dayjs | null>(
-    dayjs()
-  );
-  const [suggestedEndDate, setSuggestedEndDate] = useState<Dayjs | null>(
-    dayjs()
-  );
-
-  const [error, setError] = useState<any>(null);
-  const valid =
-    title !== "" &&
-    startDate != null &&
-    startDate.isValid() &&
-    endDate != null &&
-    endDate.isValid();
-
-  const [addChapter, isUpdating] =
-    useMutation<lecturerEditCourseChaptersMutation>(graphql`
-      mutation lecturerEditCourseChaptersMutation(
-        $chapter: CreateChapterInput!
-      ) {
-        createChapter(input: $chapter) {
-          id
-          course {
-            ...lecturerEditCourseAddChapterModalFragment
-          }
-        }
-      }
-    `);
-
-  function handleSubmit() {
-    const nextCourseNumber = course.chapters.elements.length + 1;
-
-    addChapter({
-      variables: {
-        chapter: {
-          courseId: course.id,
-          title,
-          description,
-          startDate: startDate!.toISOString(),
-          endDate: endDate!.toISOString(),
-          suggestedEndDate: suggestedEndDate?.toISOString(),
-          suggestedStartDate: suggestedStartDate?.toISOString(),
-          number: nextCourseNumber,
-        },
-      },
-      onCompleted() {
-        onClose();
-      },
-      onError(error) {
-        setError(error);
-      },
-    });
-  }
-
-  return (
-    <>
-      <Dialog maxWidth="md" open={open} onClose={onClose}>
-        <DialogTitle>Add a Chapter</DialogTitle>
-        <DialogContent>
-          {error?.source.errors.map((err: any, i: number) => (
-            <Alert key={i} severity="error" onClose={() => setError(null)}>
-              {err.message}
-            </Alert>
-          ))}
-          <Form>
-            <FormSection title="General">
-              <TextField
-                className="w-96"
-                label="Title"
-                variant="outlined"
-                value={title}
-                onChange={(e) => setTitle(e.target.value)}
-                required
-              />
-              <TextField
-                className="w-96"
-                label="Description"
-                variant="outlined"
-                value={description}
-                onChange={(e) => setDescription(e.target.value)}
-                multiline
-              />
-            </FormSection>
-
-            <FormSection title="Start and end">
-              <DatePicker
-                label="Start date"
-                value={startDate}
-                maxDate={endDate ?? undefined}
-                onChange={setStartDate}
-                slotProps={{
-                  textField: {
-                    required: true,
-                    error: startDate?.isValid() === false,
-                  },
-                }}
-              />
-              <DatePicker
-                label="End date"
-                value={endDate}
-                minDate={startDate ?? undefined}
-                defaultCalendarMonth={startDate ?? undefined}
-                onChange={setEndDate}
-                slotProps={{
-                  textField: {
-                    required: true,
-                    error: endDate?.isValid() === false,
-                  },
-                }}
-              />
-            </FormSection>
-
-            <FormSection title="Suggested start and end">
-              <DatePicker
-                label="Suggested start date"
-                value={suggestedStartDate}
-                maxDate={endDate ?? undefined}
-                onChange={setSuggestedStartDate}
-                slotProps={{
-                  textField: {
-                    required: true,
-                    error: suggestedStartDate?.isValid() === false,
-                  },
-                }}
-              />
-              <DatePicker
-                label="Suggested end date"
-                value={suggestedEndDate}
-                minDate={suggestedStartDate ?? undefined}
-                defaultCalendarMonth={suggestedStartDate ?? undefined}
-                onChange={setSuggestedEndDate}
-                slotProps={{
-                  textField: {
-                    required: true,
-                    error: suggestedEndDate?.isValid() === false,
-                  },
-                }}
-              />
-            </FormSection>
-          </Form>
-          <Backdrop open={isUpdating} sx={{ zIndex: "modal" }}>
-            <CircularProgress />
-          </Backdrop>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={onClose}>Cancel</Button>
-          <Button disabled={!valid} onClick={handleSubmit}>
-            Add
-          </Button>
-        </DialogActions>
-      </Dialog>
-      <Backdrop open={isUpdating} sx={{ zIndex: "modal" }}>
-        <CircularProgress />
-      </Backdrop>
-    </>
   );
 }
 
@@ -801,121 +195,25 @@ function AddMediaButton({
   );
 }
 
-function AddFlashcardSetModal({
-  onClose,
-  chapterId,
-  chapterRecordId,
+function AddFlashcardSetButton({
+  _chapter,
 }: {
-  onClose: () => void;
-  chapterId: string;
-  chapterRecordId: string;
+  _chapter: AddFlashcardSetModalFragment$key;
 }) {
-  const [metadata, setMetadata] = useState<ContentMetadataPayload | null>(null);
-  const [assessmentMetadata, setAssessmentMetadata] =
-    useState<AssessmentMetadataPayload | null>(null);
-  const [error, setError] = useState<any>(null);
-  const valid = metadata != null && assessmentMetadata != null;
-
-  const [createAssessment, isCreatingAssessment] =
-    useMutation<lecturerCreateFlashcardAssessmentMutation>(graphql`
-      mutation lecturerCreateFlashcardAssessmentMutation(
-        $assessment: CreateAssessmentInput!
-      ) {
-        createAssessment(input: $assessment) {
-          __id
-          __typename
-
-          ...ContentFlashcardFragment
-
-          id
-          userProgressData {
-            nextLearnDate
-          }
-        }
-      }
-    `);
-  const [createFlashcardSet, isCreatingFlashcardSet] =
-    useMutation<lecturerCreateFlashcardSetMutation>(graphql`
-      mutation lecturerCreateFlashcardSetMutation($assessmentId: UUID!) {
-        createFlashcardSet(
-          input: { assessmentId: $assessmentId, flashcards: [] }
-        ) {
-          assessmentId
-        }
-      }
-    `);
-  const isUpdating = isCreatingAssessment || isCreatingFlashcardSet;
-
-  function handleSubmit() {
-    const assessment = {
-      metadata: {
-        ...metadata!,
-        chapterId,
-        tagNames: [],
-        type: "FLASHCARDS" as ContentType,
-      },
-      assessmentMetadata: {
-        ...assessmentMetadata!,
-        skillType: assessmentMetadata!.skillType as SkillType,
-      },
-    };
-    createAssessment({
-      variables: { assessment },
-      onError: setError,
-      onCompleted(response) {
-        createFlashcardSet({
-          variables: {
-            assessmentId: response.createAssessment.id,
-          },
-          onError: setError,
-          updater(store) {
-            // Get record of chapter and of the new assignment
-            const chapterRecord = store.get(chapterRecordId);
-            const newRecord = store.get(response!.createAssessment.__id);
-            if (!chapterRecord || !newRecord) return;
-
-            // Update the linked records of the chapter contents
-            const contentRecords =
-              chapterRecord.getLinkedRecords("contents") ?? [];
-            chapterRecord.setLinkedRecords(
-              [...contentRecords, newRecord],
-              "contents"
-            );
-          },
-          onCompleted() {
-            onClose();
-          },
-        });
-      },
-    });
-    onClose();
-  }
+  const [openModal, setOpenModal] = useState(false);
 
   return (
     <>
-      <Dialog maxWidth="md" open onClose={onClose}>
-        <DialogTitle>Add flashcard set</DialogTitle>
-        <DialogContent>
-          {error?.source.errors.map((err: any, i: number) => (
-            <Alert key={i} severity="error" onClose={() => setError(null)}>
-              {err.message}
-            </Alert>
-          ))}
-          <Form>
-            <ContentMetadataFormSection onChange={setMetadata} />
-            <AssessmentMetadataFormSection onChange={setAssessmentMetadata} />
-          </Form>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={onClose}>Cancel</Button>
-          <Button disabled={!valid} onClick={handleSubmit}>
-            Save
-          </Button>
-        </DialogActions>
-      </Dialog>
-      <Backdrop open={isUpdating} sx={{ zIndex: "modal" }}>
-        <CircularProgress />
-      </Backdrop>
+      <Button startIcon={<Add />} onClick={() => setOpenModal(true)}>
+        Add flashcards
+      </Button>
+
+      {openModal && (
+        <AddFlashcardSetModal
+          onClose={() => setOpenModal(false)}
+          _chapter={_chapter}
+        />
+      )}
     </>
   );
 }
